Update version to 0.4.0
Fix powershell install on GH action

Set gh action to at latest

Set gh action to fix error

Set gh action to v2

remove powershell install action

Install powershell manually in builder

Update Node version for builder

Manual install PowerShell on macOS; separate from Linux

Fix error in macOS install script

Force link creation in PowerShell install

Install Xcode Command Line Tools onto builder

Install vcpkg dependencies for builder

Run vcpkg using env var and remove pwsh workflows

Build setup continues on error

Debugging build-setup.ps1

Trying build on macos 14

Increase timeout minutes for builder

Divide up environment settings by OS

Remove fail-fast strategy and ubuntu build

Try again to remove ubuntu from builder

Separate vcpkg per platform and cache build

Add binarysource to vcpkg

Remove unofficial-libuv from macos build

Undo fail-fast directive

Remove unofficial-libuv from Apple build

Cache npm deps. Fix Electron Forge runner

Fix Electron Forge platform for macos

Fix macOS universal duplicate resources error

Downgrade Electron Forge to fix universal package

Build Intel and Apple arch separately

Remove duplicate EF make command

Fix artifact paths. Add explicit VCPKG_ROOT env var

Update wingman.cpp

Fix EF build platform for Windows

Fix build ux script.

Set default Shell to bash

Fix for macos cannot find appdmg

Fix package.json folder location for jq

+++ ux/build-ux-ci.ps1
@@ -1,19 +1,26 @@
 param(
     [Parameter(Mandatory = $true)]
     [ValidateSet("windows", "linux", "macos")]
-    [string]$BuildPlatform
+    [string]$BuildPlatform,
+
+    [Parameter(Mandatory = $false)]
+    [switch]$Force
 )
 
 try {
     # Clean up the previous build artifacts
-    Write-Host "Cleaning previous build artifacts..."
-    Remove-Item -Recurse -Force -ErrorAction SilentlyContinue node_modules
-    Remove-Item -Recurse -Force -ErrorAction SilentlyContinue out
-    Remove-Item -Recurse -Force -ErrorAction SilentlyContinue .next
+
+    if ($Force) {
+        Write-Host "Cleaning previous build artifacts..."
+        Remove-Item -Recurse -Force -ErrorAction SilentlyContinue node_modules
+        Remove-Item -Recurse -Force -ErrorAction SilentlyContinue out
+        Remove-Item -Recurse -Force -ErrorAction SilentlyContinue .next
+    }
 
     # Install dependencies
-    Write-Host "Installing dependencies..."
-    npm ci --cache .npm --prefer-offline
+    Write-Host "Installing Node dependencies..."
+    # npm ci --cache .npm --prefer-offline
+    npm install
     if ($LASTEXITCODE -ne 0) {
         throw "npm install failed" 
     }
@@ -30,15 +37,18 @@ try {
     Copy-Item ".next/static" ".next/standalone/.next" -Recurse -Force
 
     # Determine architecture based on platform
-    $arch = switch ($BuildPlatform) {
+    $arch = ""
+    $platform = ""
+    switch ($BuildPlatform) {
         "windows" {
-            "x64" 
+            $arch = "x64"
+            $platform = "win32"
         }
         "linux" {
-            "x64" 
+            $arch = "x64"
+            $platform = "linux"
         }
         "macos" {
-            "universal" 
         } # For universal macOS build
         default {
             throw "Unsupported platform: $BuildPlatform" 
@@ -47,7 +57,13 @@ try {
 
     # Build the Electron app
     Write-Host "Building Electron app..."
-    electron-forge make --platform=$BuildPlatform --arch=$arch
+    if ($BuildPlatform -eq "macos") {
+        # build x64 and arm64
+        ./node_modules/.bin/electron-forge make --platform=darwin --arch=x64
+        ./node_modules/.bin/electron-forge make --platform=darwin --arch=arm64
+    } else {
+        ./node_modules/.bin/electron-forge make --platform=$platform --arch=$arch 
+    }
     if ($LASTEXITCODE -ne 0) {
         throw "electron-forge make failed" 
     }
